requirements.txt
Flask==3.0.3
gunicorn==22.0.0
openai>=1.40.0,<2
python-dotenv==1.0.1
stripe==11.6.0
requests==2.32.3
